extract repository/git-related test to own file
