+++ tests/prompt_test.py
@@ -1,10 +1,7 @@
 import os
-import tempfile
 from pathlib import Path
 
-import git
-
-from pure import prompt, colors, repository
+from pure import prompt, colors
 
 SUCCESS = 0
 FAIL = 1
@@ -58,35 +55,3 @@ def test_displays_virtual_env_when_activated():
 def test_prompt_layout():
     assert prompt.layout() == "\n{current_working_path} {git_active_branch}{git_is_dirty}\n{virtual_env}{prompt_symbol} "
 
-
-def test_contains_git_branch_name():
-    with tempfile.TemporaryDirectory() as tmp_repo:
-        empty_repo = git.Repo.init(tmp_repo)
-
-        assert 'master' in str(repository.active_branch(tmp_repo))
-
-
-def test_git_branch_name_color_is_mute():
-    with tempfile.TemporaryDirectory() as tmp_repo:
-        empty_repo = git.Repo.init(tmp_repo)
-
-        assert str(repository.active_branch(tmp_repo)) == str(colors.mute('master'))
-
-
-def test_displays_when_repo_is_dirty():
-    with tempfile.TemporaryDirectory() as tmp_repo:
-        empty_repo = git.Repo.init(tmp_repo)
-        new_file = tempfile.NamedTemporaryFile(dir=tmp_repo)
-
-        assert '*' in str(repository.is_dirty(tmp_repo))
-        new_file.close()
-
-
-def test_repo_is_dirty_color_is_mute():
-    with tempfile.TemporaryDirectory() as tmp_repo:
-        empty_repo = git.Repo.init(tmp_repo)
-        new_file = tempfile.NamedTemporaryFile(dir=tmp_repo)
-
-        assert str(repository.is_dirty(tmp_repo)) == str(colors.mute('*'))
-
-        new_file.close()
